Remove models folder and relocate classes
