+++ samples/java_springboot/54.teams-task-module/src/main/java/com/microsoft/bot/sample/teamstaskmodule/TaskModuleResponseFactory.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.microsoft.bot.sample.teamstaskmodule.models;
+package com.microsoft.bot.sample.teamstaskmodule;
 
 import com.microsoft.bot.schema.teams.TaskModuleContinueResponse;
 import com.microsoft.bot.schema.teams.TaskModuleMessageResponse;
