+++ samples/java_springboot/54.teams-task-module/src/main/java/com/microsoft/bot/sample/teamstaskmodule/TeamsTaskModuleBot.java
@@ -10,12 +10,6 @@ import com.microsoft.bot.builder.TurnContext;
 import com.microsoft.bot.builder.teams.TeamsActivityHandler;
 import com.microsoft.bot.connector.Async;
 import com.microsoft.bot.integration.Configuration;
-import com.microsoft.bot.sample.teamstaskmodule.models.AdaptiveCardTaskFetchValue;
-import com.microsoft.bot.sample.teamstaskmodule.models.CardTaskFetchValue;
-import com.microsoft.bot.sample.teamstaskmodule.models.TaskModuleIds;
-import com.microsoft.bot.sample.teamstaskmodule.models.TaskModuleResponseFactory;
-import com.microsoft.bot.sample.teamstaskmodule.models.TaskModuleUIConstants;
-import com.microsoft.bot.sample.teamstaskmodule.models.UISettings;
 import com.microsoft.bot.schema.Attachment;
 import com.microsoft.bot.schema.CardAction;
 import com.microsoft.bot.schema.HeroCard;
