--- samples/java_springboot/54.teams-task-module/src/main/java/com/microsoft/bot/sample/teamstaskmodule/models/package-info.java
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See License.txt in the project root for
-// license information.
-
-/**
- * This package contains the classes for Teams Task Module Sample.
- */
-package com.microsoft.bot.sample.teamstaskmodule.models;
